requirements-dev.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
isort==5.13.2
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.9.4
ruff~=0.5.0
pre-commit~=3.7.1
